Split mapping functions of ActivePageTable into mapper subtype
